v0.1.10 - 添加Frpc启动拦截提示 - 添加签到按钮 - 修改Frpc日志系统 - 修复错误的日志检测 - 添加更多自动错误检查 - 适配新版API值类型变动
