+++ nyalcf_core/lib/controllers/panel_controller.dart
@@ -1,6 +1,6 @@
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/utils/network/dio/other/other.dart';
+import 'package:nyalcf_core/network/dio/other/other.dart';
 
 class DPanelController extends GetxController {
   static var loaded = false;
