+++ nyalcf_core/lib/network/response_type.dart
@@ -1,0 +1,11 @@
+class Response {
+  bool status;
+  String message;
+  Map<String, dynamic> data;
+
+  Response({
+    required this.status,
+    required this.message,
+    required this.data,
+  });
+}
