+++ nyalcf_core/lib/tasks/updater.dart
@@ -5,7 +5,7 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:nyalcf_core/models/update_info_model.dart';
 import 'package:nyalcf_core/tasks/basic.dart';
 import 'package:nyalcf_core/utils/logger.dart';
-import 'package:nyalcf_core/utils/network/dio/launcher/launcher.dart';
+import 'package:nyalcf_core/network/dio/launcher/launcher.dart';
 import 'package:nyalcf_core/utils/universe.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
