--- nyalcf_core/lib/utils/network/dio/frpc/frpc.dart
@@ -1,4 +1,0 @@
-library frpc;
-
-export 'download_frpc.dart';
-export 'version_frpc.dart';
