--- nyalcf_core/lib/utils/network/dio/launcher/launcher.dart
@@ -1,3 +1,0 @@
-library launcher;
-
-export 'update_launcher.dart';
