--- nyalcf_core/lib/utils/network/dio/other/other.dart
@@ -1,4 +1,0 @@
-library other;
-
-export 'announcement_other.dart';
-export 'auto_sign_other.dart';
