--- nyalcf_core/lib/utils/network/dio/proxies/proxies.dart
@@ -1,5 +1,0 @@
-library proxies;
-
-export 'configuration_proxies.dart';
-export 'get_proxies.dart';
-export 'status_proxies.dart';
