--- nyalcf_core/lib/utils/network/response_type.dart
@@ -1,11 +1,0 @@
-class Response {
-  bool status;
-  String message;
-  Map<String, dynamic> data;
-
-  Response({
-    required this.status,
-    required this.message,
-    required this.data,
-  });
-}
