+++ nyalcf_core/lib/utils/proxies_getter.dart
@@ -5,7 +5,7 @@ import 'package:nyalcf_core/models/user_info_model.dart';
 import 'package:nyalcf_core/storages/prefs/user_info_prefs.dart';
 import 'package:nyalcf_core/storages/stores/proxies_storage.dart';
 import 'package:nyalcf_core/utils/logger.dart';
-import 'package:nyalcf_core/utils/network/dio/proxies/proxies.dart';
+import 'package:nyalcf_core/network/dio/proxies/proxies.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 
 class ProxiesGetter {
