+++ nyalcf_ui/lib/models/frpc_download_dialog.dart
@@ -4,7 +4,7 @@ import 'package:get/get.dart';
 import 'package:nyalcf_core/controllers/frpc_setting_controller.dart';
 import 'package:nyalcf_core/storages/configurations/frpc_configuration_storage.dart';
 import 'package:nyalcf_core/utils/logger.dart';
-import 'package:nyalcf_core/utils/network/dio/frpc/frpc.dart';
+import 'package:nyalcf_core/network/dio/frpc/frpc.dart';
 
 class FrpcDownloadDialogX {
   FrpcDownloadDialogX({required this.context});
