+++ nyalcf_ui/lib/views/auth/login.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import 'package:nyalcf_core/storages/prefs/user_info_prefs.dart';
-import 'package:nyalcf_core/utils/network/dio/auth/auth.dart';
+import 'package:nyalcf_core/network/dio/auth/auth.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
