+++ pubspec.yaml
@@ -16,7 +16,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 0.1.9+4
+version: 0.1.10
 
 environment:
   sdk: '>=3.1.5 <4.0.0'
@@ -75,8 +75,6 @@ flutter:
   assets:
     - icon.png
     - icon.ico
-    - runtime/VC_redist.x64.exe
-    - runtime/VC_redist.x86.exe
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
